Allow for different shells to be chosen for Command execution
I changed the Django Command model to use models.choices.
At the moment this will only enforce the default shell (cmd, sh),
PowerShell, and Bash.

This is also reflected client-side in an unordered map with supported
shells and the text that has to be prepended and appended to the command
text for that command to execute properly.

Relevant unit tests were added.